wpt_babel/content-security-policy/style-src/style-src-inline-style-nonce-allowed.html

<!DOCTYPE html>
<html>
 <head>
  <meta content="style-src 'nonce-nonceynonce';" http-equiv="Content-Security-Policy"/>
  <script src="/resources/testharness.js">
  </script>
  <script src="/resources/testharnessreport.js">
  </script>
  <script>
   "use strict";

var t = async_test("Style with correct nonce should load");
document.addEventListener("securitypolicyviolation", t.unreached_func("Should not trigger a security policy violation"));
  </script>
  <style nonce="nonceynonce">
   #content {
            margin-left: 2px;
        }
  </style>
 </head>
 <body>
  <div id="log">
  </div>
  <div id="content">
   Lorem ipsum
  </div>
  <script>
   "use strict";

t.step(function () {
  var contentEl = document.getElementById("content");
  var marginLeftVal = getComputedStyle(contentEl).getPropertyValue('margin-left');
  assert_equals(marginLeftVal, "2px");
  t.done();
});
  </script>
 </body>
</html>
